feat: move read preview user in empty wallet
